requirements.txt
discord.py
boto3
flask
requests